refactor(response_cache): separate cache key in their own structs
